refactor: reduce visibility of modules and types in profit
