+++ protocol/contracts/profit/src/typedefs.rs
@@ -1,1 +1,1 @@
-pub type CadenceHours = u16;
+
